frontend/requirements_flask.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# 根據你的 RAG 系統可能需要的其他套件
# 請根據你的 ai_doctor 目錄中的 requirements.txt 添加其他依賴

# 核心依賴
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.16
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
chromadb>=0.4.22

# 文檔處理
langchain-text-splitters>=0.0.1
pypdf>=3.17.0

# Web應用
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2

# 工具依賴
numpy>=1.24.0
typing-extensions>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1
pandas